[Refactor] Socket.io 리팩토링(1) 콘솔로그 제거 및, 중복코드 개선
