+++ src/pages/chat/ChatRoom.tsx
@@ -1,34 +1,25 @@
 import { FC } from 'react'
-import { useSocket } from '../../hooks'
+import * as Hook from '../../hooks'
+import { styled } from 'styled-components'
 
 
 export const ChatRoom: FC = () => {
-  const{
+  const {
     // 채팅부분
-    sendMsg, onChangeInput, onSendMsg, onLeaveRoom,
+    sendMsg,
+    onChangeInput,
+    onSendMsg,
+    onLeaveRoom,
     // WecRTC 부분
     showWebRTC,
     peerAVideoRef,
     peerBVideoRef,
-    onToggleWebRTC,
-    onMute, 
     mute,
-    onCamera,
     camara,
-    audioList,
-    onChangeAudio
-  } = useSocket()
-
-  // console.log(audioList)
-
-  /*
-  Record<string, string>
-  audioList
-  deviceId: "54D9354A71B935C33CDF4848B383E281077342EB"  
-  groupId: "25560712DC2F13C5CF3BF5EC5F3EB586FDB2183A"
-  kind: "audioinput"
-  label: "MacBook Air 마이크"
-  */
+    onToggleWebRTC,
+    onMute,
+    onCamera,
+  } = Hook.useSocketRoom()
 
   return (
     <div>
@@ -44,25 +35,17 @@ export const ChatRoom: FC = () => {
         <h1>WebRTC</h1>
         <button onClick={onCamera}>카메라 {!camara ? "끄기" : "켜기"}</button>
         <button onClick={onMute}>음소거 {!mute ? "하기" : "끄기"}</button>
-        <select onChange={onChangeAudio}>
-          {audioList.map(({deviceId, label}:Record<string, string>) => <option key={deviceId} value={deviceId} children={label}/>)}
-        </select>
-        <video ref={peerAVideoRef} style={{
-        width: 1000,
-        height: 800,
-        backgroundColor: "black",
-        transform: "scaleX(-1)"
-      }}
-        autoPlay />
-        <hr/>
-        <video ref={peerBVideoRef} style={{
-        width: 1000,
-        height: 800,
-        backgroundColor: "black",
-        transform: "scaleX(-1)"
-      }}
-        autoPlay />
+        <Video ref={peerAVideoRef} autoPlay />
+        <hr />
+        <Video ref={peerBVideoRef} autoPlay />
       </div>}
     </div>
   )
 }
+
+const Video = styled.video`
+  width: 1000;
+  height: 800;
+  background-color: black;
+  transform: scaleX(-1);
+`
